feat: #WB2-996, do not mix `[]` and `null` in ExplorerMessage.rights because they have different meaning
An empty array was used as a default value for the field `rights` of ExplorerMessage but it caused shares to be reset when the creator of a resource updated the resource properties (`null` should have been used instead).
So we now clearly distinguish between `null` and `[]` values while setting `rights`.